% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{calculate_earnings_loop}
\alias{calculate_earnings_loop}
\title{Calculate earnings}
\usage{
calculate_earnings_loop(data, var_price, verbose)
}
\arguments{
\item{data}{tibble with stock data (see kb.yahoo::load_data)}

\item{var_price}{character string with price column (options: "Open", "High", "Low", "Close", "Adjusted" or "Avg")}

\item{verbose}{logical indicating if loop counter should be on or not}
}
\value{

}
\description{
Calculate earnings
}
