% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{cumsum_with_reset}
\alias{cumsum_with_reset}
\title{Cumulative sum with reset}
\usage{
cumsum_with_reset(x, reset)
}
\arguments{
\item{x}{numeric vector to sum}

\item{reset}{logical vector used to reset the cumulative sum (for instance is.na(x))}
}
\value{
numeric vector
}
\description{
Cumulative sum with reset
}
