% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{data_prep_avg_buy_price}
\alias{data_prep_avg_buy_price}
\title{Data prep average buy price}
\usage{
data_prep_avg_buy_price(data)
}
\arguments{
\item{data}{tibble obtained from data_prep_stock_holdings}
}
\value{
tibble with column AvgBuyPrice (addition to input columns)
}
\description{
Data prep average buy price
}
