% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{data_prep_stock_holdings}
\alias{data_prep_stock_holdings}
\title{Data prep stock holdings earnings}
\usage{
data_prep_stock_holdings(data, var_price)
}
\arguments{
\item{data}{tibble with stock data (see kb.yahoo::load_data)}

\item{var_price}{character string with price column (options: "Open", "High", "Low", "Close", "Adjusted" or "Avg")}
}
\value{
tibble with additional columns, including stockHoldings
}
\description{
Data prep stock holdings earnings
}
