% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{get_all_actions}
\alias{get_all_actions}
\title{Get all actions}
\usage{
get_all_actions()
}
\value{
character vector with possible actions
}
\description{
Get all actions
}
\examples{
get_all_actions()
}
