% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{get_avg_buy_price}
\alias{get_avg_buy_price}
\title{Get average buy price}
\usage{
get_avg_buy_price(StocksHolding, BuyPrice, BuyAmount)
}
\arguments{
\item{StocksHolding}{tibble columns defined in calculate_earnings}
}
\description{
Get average buy price
}
\details{
This is a helper function for calculate_earnings
}
