% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{get_avg_buy_price_loop}
\alias{get_avg_buy_price_loop}
\title{Get average buy price}
\usage{
get_avg_buy_price_loop(n_stocks_cur, n_stocks_new, price_cur, price_new)
}
\arguments{
\item{n_stocks_cur}{number of stocks currently holding (numeric value). Use NA if currently not holding any}

\item{n_stocks_new}{number of stocks buying (numeric value)}

\item{price_cur}{current average buy price (numeric value). Use NA if currently not holding any}

\item{price_new}{new buy price}
}
\value{
numeric value
}
\description{
Get average buy price
}
\details{
get_avg_buy_price updates the average buy price by averaging current
buy price and new buy price weighted by the number of stocks having and buying
}
\examples{
get_avg_buy_price(n_stocks_cur = 1, n_stocks_new = 1, price_cur = 15, price_new = 5)
}
