% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_lag_diff.R
\name{get_price_lag_diff}
\alias{get_price_lag_diff}
\title{Get price lag diff}
\usage{
get_price_lag_diff(data, price_var)
}
\arguments{
\item{data}{tibble with stock data}

\item{price_var}{price variable to use (Close, Low, High or Open)}
}
\value{
tibble with additional columns
}
\description{
Get price lag diff
}
