% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_lag_diff.R
\name{get_price_optimum}
\alias{get_price_optimum}
\title{Get price optimum}
\usage{
get_price_optimum(data)
}
\arguments{
\item{data}{tibble obtained from get_price_lag_diff}
}
\value{
tibble with additional columns
}
\description{
Get price optimum
}
