% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{get_relative_diff}
\alias{get_relative_diff}
\title{Get relative difference}
\usage{
get_relative_diff(x, y, digits = 2)
}
\arguments{
\item{x}{numeric vector (benchmark value)}

\item{y}{numeric vector (comparison value)}

\item{digits}{number of digits (default is 2)}
}
\description{
Get relative difference
}
